Make custom events immutable
Create custom event from the copy of the original event.
This makes possible for example to stop the propagation of tap event,
but keep the propagaton of doubletab.
Before stopping the propagaton on one event stop it for all,
because the events share the same event object.
Currently this uses es6 features.